Create index.0 based on all packages in packages dir
This no longer uses only the copy in $PKGFLAT (.all directory), but
instead fetches the names from the subdirs in the packages dir. Right
now this is not an improvement, but it's a preparation so that the .all
directory can be removed and replaced with a redirection in the
webserver.

To prevent an existing .all directory from duplicating the results, the
list is piped through `sort -u`. This can probably be removed again once
the .all directory is removed.

Signed-off-by: Florian Pritz <bluewind@xinu.at>